updates to biller and chat protocol docs.
added protutil.{c,h}, which is a bunch of routines factored out of
chatnet and the new biller module.

modified chatnet to use protutil.

fixed ship type on entering (from snrrrub).

added authenticated flag for each player and made capman only set groups
based on name for authenticated players.

added authenticated flag to authdata struct and changed all existing
auth modules to set authenticated when appropriate.

removed old billcore code and added new biller module.

made chat pass around message sounds all over the place, and only allow
sounds for players with the sendsoundmessages capability.

added server-interla remote messages to chat (finally!).

made dontlogparams commands hard-coded instead of in config file
(suggested by priitk).

cleaned up logging statements in LoadMod.

fixed bug in objects (i think priitk found this one).

updated turf stuff to 3.6 (from gigakiller).

added readlvl testing utility (it draws pictures of .lvls in ascii),
among other things.

+++ src/mapdata.c
@@ -124,19 +124,19 @@ EXPORT int MM_mapdata(int action, Imodman *_mm, Arena *arena)
 
 int read_lvl(char *name, struct MapData *md)
 {
-	unsigned short bm;
+	u16 bm;
 	struct TileData td;
 	int flags = 0, errors = 0;
 	sparse_arr arr;
 
-	FILE *f = fopen(name,"r");
+	FILE *f = fopen(name, "r");
 	if (!f) return 1;
 
 	/* first try to skip over bmp header */
 	fread(&bm, sizeof(bm), 1, f);
 	if (bm == 0x4D42)
 	{
-		unsigned long len;
+		u32 len;
 		fread(&len, sizeof(len), 1, f);
 		fseek(f, len, SEEK_SET);
 	}
